AI-based-Image-Analysis/project_root/requirements.txt
streamlit
Pillow
torchvision
torch
matplotlib
numpy
easyocr
transformers
json
unittest
transformers
numpy
zipfile